requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.5.0
django-ckeditor==5.8.0
django-configurations==2.1
django-debug-toolbar==2.0
django-extensions==2.2.5
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.0
mccabe==0.6.1
Pillow==7.0.0
progressbar2==3.47.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-utils==2.3.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
Unidecode==1.1.1
